backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.12
alembic==1.14.0
Werkzeug==3.1.3
requests==2.31.0
